Move siteid package to sourcegraph/internal
