+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -233,7 +233,6 @@ go_library(
         "//cmd/frontend/internal/app/updatecheck",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
-        "//cmd/frontend/internal/siteid",
         "//enterprise/cmd/worker/shared/sourcegraphoperator",
         "//internal/actor",
         "//internal/adminanalytics",
@@ -308,6 +307,7 @@ go_library(
         "//internal/search/zoekt",
         "//internal/service/servegit",
         "//internal/settings",
+        "//internal/siteid",
         "//internal/src-prometheus",
         "//internal/suspiciousnames",
         "//internal/symbols",
