+++ cmd/frontend/internal/app/jscontext/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/hooks",
         "//cmd/frontend/internal/app/assetsutil",
-        "//cmd/frontend/internal/siteid",
         "//cmd/frontend/webhooks",
         "//internal/actor",
         "//internal/auth/providers",
@@ -25,6 +24,7 @@ go_library(
         "//internal/env",
         "//internal/insights",
         "//internal/lazyregexp",
+        "//internal/siteid",
         "//internal/types",
         "//internal/version",
         "//schema",
