+++ cmd/frontend/internal/app/ui/BUILD.bazel
@@ -88,7 +88,6 @@ go_test(
         "//cmd/frontend/envvar",
         "//cmd/frontend/globals",
         "//cmd/frontend/internal/app/ui/router",
-        "//cmd/frontend/internal/siteid",
         "//internal/api",
         "//internal/authz",
         "//internal/conf",
@@ -100,6 +99,7 @@ go_test(
         "//internal/gitserver/gitdomain",
         "//internal/repoupdater",
         "//internal/search/result",
+        "//internal/siteid",
         "//internal/types",
         "//internal/version",
         "//schema",
