+++ cmd/frontend/internal/app/updatecheck/BUILD.bazel
@@ -16,7 +16,6 @@ go_library(
         "//cmd/frontend/globals",
         "//cmd/frontend/hubspot",
         "//cmd/frontend/hubspot/hubspotutil",
-        "//cmd/frontend/internal/siteid",
         "//internal/conf",
         "//internal/conf/deploy",
         "//internal/database",
@@ -28,6 +27,7 @@ go_library(
         "//internal/metrics",
         "//internal/pubsub",
         "//internal/redispool",
+        "//internal/siteid",
         "//internal/types",
         "//internal/usagestats",
         "//internal/version",
