+++ cmd/frontend/internal/cli/BUILD.bazel
@@ -32,7 +32,6 @@ go_library(
         "//cmd/frontend/internal/cli/middleware",
         "//cmd/frontend/internal/httpapi",
         "//cmd/frontend/internal/httpapi/router",
-        "//cmd/frontend/internal/siteid",
         "//cmd/frontend/oneclickexport",
         "//internal/actor",
         "//internal/adminanalytics",
@@ -74,6 +73,7 @@ go_library(
         "//internal/requestclient",
         "//internal/service",
         "//internal/session",
+        "//internal/siteid",
         "//internal/symbols",
         "//internal/sysreq",
         "//internal/trace",
