+++ internal/usagestats/BUILD.bazel
@@ -35,7 +35,6 @@ go_library(
     visibility = ["//:__subpackages__"],
     deps = [
         "//cmd/frontend/envvar",
-        "//cmd/frontend/internal/siteid",
         "//internal/actor",
         "//internal/conf",
         "//internal/database",
@@ -46,6 +45,7 @@ go_library(
         "//internal/pubsub",
         "//internal/redispool",
         "//internal/search",
+        "//internal/siteid",
         "//internal/timeutil",
         "//internal/types",
         "//internal/version",
